// src/app/shared/models/nesting-algo.model.ts
export class NestingAlgo {
  maximumSheetLength: number;
  ierationNumber: string;
  sheetWidth: number;
  boundaryBuffer: number;
  multiPartsSystem: boolean;
  image: boolean;
  scale: boolean;
  partInfoId: number;
  patternBased: boolean;
  softTooling: boolean;
  quickSoftTooling: boolean;
  rectangularHeight: number;
  rectangularWidth: number;
  imageValue: string;
  margin: number;
  numberOfLanes: number;
  thickness: number;
  partAllowance: number;
  edgeAllowance: number;
  bottomClamping: number;
  leftClamping: number;
}